SeedKeeper - add exportAuthentikey() command
Compared to getBIP32AuthentiKey(), this method returns the Authentikey
even if the card is not seeded.
For SeedKeeper encrypted seed import, we use the authentikey as a
Trusted Pubkey for the ECDH key exchange, thus the authentikey must
be available before the Satochip is seeded. Before a seed is available,
the authentiey is generated oncard randomly in the constructor.

(Work In Progress, still more testing required)