SeedKeeper v0.2: add BIP32 support with getBIP32ExtendedKey()
The function computes the Bip32 extended key derived from the master key and returns either the
32-bytes x-coordinate of the public key, or the 32-bytes private key, signed by the authentikey.
The Path for the derivation is provided in the apdu data.

The function is mostly compatible with Satochip, with the addition of sid (secret_id of the masterseed) appended in the apdu data
P2 parameters can include additional option masks:
*0x01: if set, use secure export (currently not supported!), otherwise use plain export
*0x02: if set, return privkey bytes, else public key (todo!)
*0x04: (RFU) if set, add final BIP85 HMAC derivation (currently not supported!)

In this current implementation, none of these options is supported!