SeedKeeper v0.2: importSecret() and exportSecret() memory optimization
Previously, recvBuffer was used as a temporary buffer to store the full secret during import/export steps.
Hence, recvBuffer size was a limiting factor when importing/exporting a secret.

We remove the need to make a full copy of the secret to recvBuffer during import/export.
Instead, we copy data directly from secret storage (om_secrets) in chunks.

This allows to import/export secrets without size limitation (appart from the inherent limitation of storage capacity).

For importSecret(), we need to provide the exact size of memory needed to store the (encrypted) secret in memory since memory is allocated in advance.
This must be provided by the client in the init() phase. The size must take padding into acount.
This size will be used to create object before any secret data is provided.
If size is too small, an exception will be thrown. If size is too big, object will be clamped after import.