SeedKeeper v0.2: add exportSecretToSatochip() method
This function exports a secret for secure import to a Satochip.
The secret is encrypted with a key is generated with ECDH, using the Satochip authentikey.
Compared to the exportSecret() method, this method is executed in one phase, and the secret size is reduced to the minimum (max 64b).
Only Masterseeds & 2FA secrets can be exported this way.

     * ins: 0xA8
     * p1: 0x00
     * p2: 0x00
     * data: [ id(2b) | id_pubkey(2b) ]
     * return: [ id(2b) | header(13b) | IV(16b) | encrypted_secret_size(2b) | encrypted_secret | hmac_size(2b) | hmac(20b) ]